KNOX-2344 - Renamed 'hadoopxml' to 'hadoop-xml'
