+++ gateway-release/pom.xml
@@ -439,7 +439,7 @@
         </dependency>
         <dependency>
             <groupId>org.apache.knox</groupId>
-            <artifactId>gateway-topology-hadoopxml</artifactId>
+            <artifactId>gateway-topology-hadoop-xml</artifactId>
         </dependency>
         <dependency>
             <groupId>org.apache.knox</groupId>
