+++ gateway-server/pom.xml
@@ -67,7 +67,7 @@
         </dependency>
         <dependency>
             <groupId>org.apache.knox</groupId>
-            <artifactId>gateway-topology-hadoopxml</artifactId>
+            <artifactId>gateway-topology-hadoop-xml</artifactId>
         </dependency>
         <dependency>
             <groupId>org.apache.knox</groupId>
