+++ gateway-server/src/main/java/org/apache/knox/gateway/GatewayServer.java
@@ -48,8 +48,8 @@ import org.apache.knox.gateway.topology.Topology;
 import org.apache.knox.gateway.topology.TopologyEvent;
 import org.apache.knox.gateway.topology.TopologyListener;
 import org.apache.knox.gateway.topology.discovery.advanced.AdvancedServiceDiscoveryConfigurationMonitor;
-import org.apache.knox.gateway.topology.hadoopxml.HadoopXmlResourceMonitor;
-import org.apache.knox.gateway.topology.hadoopxml.HadoopXmlResourceParser;
+import org.apache.knox.gateway.topology.hadoop.xml.HadoopXmlResourceMonitor;
+import org.apache.knox.gateway.topology.hadoop.xml.HadoopXmlResourceParser;
 import org.apache.knox.gateway.trace.AccessHandler;
 import org.apache.knox.gateway.trace.KnoxErrorHandler;
 import org.apache.knox.gateway.trace.TraceHandler;
