+++ gateway-topology-hadoop-xml/src/test/resources/testDescriptor.xml
