+++ gateway-topology-hadoop-xml/src/test/resources/testDescriptorConfigurationWithNonHadoopStyleConfiguration.xml
