+++ gateway-topology-hadoop-xml/src/test/resources/testDescriptorConfigurationWithWrongDescriptor.xml
