+++ gateway-topology-hadoop-xml/src/test/resources/testDescriptorWithAdminProviderConfigRemovedUserDnTemplate.xml
