+++ gateway-topology-hadoop-xml/src/test/resources/testDescriptorWithAdminProviderConfigUpdatedLdapUrl.xml
