+++ gateway-topology-hadoop-xml/src/test/resources/testDescriptorWithoutDiscoveryDetails.xml
